UI stuff renamed to `gui`
