+++ CHANGELOG.md
@@ -12,7 +12,7 @@ All notable changes to this project will be documented in this file.
 
 - Switched to `hatching` as a build system, now correct install optional dependencies.
 - Renamed methods, attributes that was `shadowing a Python builtins`. Enabled additional `Ruff` linters checks.
-- Regroup APIs, now Users related stuff starts with `user`, file related stuff with `file`.
+- Regroup APIs, now Users related stuff starts with `user`, file related stuff with `file`, UI stuff with `gui`.
 
 ## [0.0.26 - 2023-07-29]
