+++ tests/appcfg_prefs_ex_test.py
